First Commit
This is my first release of AlexaIOT which will allow your raspberry pi
2 or 3 (developed on 3) that's running Windows IOT to use the Alexa
Voice Service. This is primarily a proof-of-concept idea because I've
seen examples for Java and python but not much for .NET this is my trial
and error way of contributing to the open source community in a way..
BEWARE - Bugs ahead.

I do not guarantee this will work out of box if something is broken
please let me know and I will try to make it compatible for everyone and
if you found a way to fix or improve anything PLEASE share it.