Major update
Pretty big update. Downchannel is now successfully created and can now
receive directives from Alexa. started work on Timers API and still
trying to figure out the direction of this project.

Some more minor changes of course. Recording is now sent to a stream
instead of being recorded to a file potentially causes a dead lock with
the files.  I need to find a way now to play audio from a stream but Im
currently limited to audio graph because my USB mic is detected as a
speaker which is not the case.